streamlit_app/requirements.txt
streamlit
psycopg2-binary
pandas
plotly
reportlab
bcrypt
